Improve the test files organization
